+++ tutorials/30-communication-and-bbs/20-middleware/senderconst.cpp
@@ -13,7 +13,7 @@ public:
 	MyPublisherC() : StaticThread<>("SenderConst") { }
 	void run () {
 		TIME_LOOP(3*SECONDS, 3*SECONDS) {
-			PRINTF(CONST "%lld\n", NOW());
+			PRINTF(CONST "%lld\n", static_cast<long long>(NOW()));
 			PRINTF(CNT2 "%ld\n", CNT);
 			counter2.publish(CNT);
 		}
