+++ tutorials/30-communication-and-bbs/40-middleware-distributed/receiver11.cpp
@@ -8,7 +8,7 @@ struct JustPrint11 : public Subscriber {
     JustPrint11() : Subscriber(counter1, "justprint11") { }
     long put([[gnu::unused]] const long topicId, [[gnu::unused]] const long len, void* msg, [[gnu::unused]] const NetMsgInfo& netMsgInfo) {
         PRINTF("%s%ld",  R11, *(long*)msg);
-        PRINTF("%s%lld", MSGCNT, getNumberOfReceivedMsgsFromNetwork());
+        PRINTF("%s%lld", MSGCNT, static_cast<long long>(getNumberOfReceivedMsgsFromNetwork()));
         return true;
     }
 } justPrint11;
