+++ tutorials/30-communication-and-bbs/80-gateway/receiver1.cpp
@@ -6,7 +6,7 @@ struct JustPrint1 : public Subscriber {
     long put(const long topicId, const long len, void* msg, const NetMsgInfo& netMsgInfo) {
         PRINTF("at %9.3f       counter1: got (topic, len, mesg, netinfo): %ld, %ld, %d, (%d, %lld, %d, %d)\n", SECONDS_NOW(),
                topicId, len, *(int32_t*)msg,
-               netMsgInfo.senderNode, netMsgInfo.sentTime, netMsgInfo.senderThreadId, netMsgInfo.linkId);
+               netMsgInfo.senderNode, static_cast<long long>(netMsgInfo.sentTime), netMsgInfo.senderThreadId, netMsgInfo.linkId);
         return true;
     }
 } justPrint1;
