+++ tutorials/60-tutorial-big-application/simple/app03/app03.cpp
@@ -10,7 +10,7 @@ public:
     AppT() : StaticThread<>("AppT03:", 400)  { }
     void run() {
         TIME_LOOP(1*SECONDS, 1*SECONDS) {
-            PRINTF("APP3:  counter2: %d\n", Common::counter2);
+            PRINTF("APP3:  counter2: %d\n", static_cast<int>(Common::counter2));
         } // loop
     }
